refactor: refactor new endpoint code into correct files
