--- src/app/modules/form/admin-form/fields/fields.errors.ts
@@ -1,7 +1,0 @@
-import { ApplicationError } from '../../../core/core.errors'
-
-export class EditFieldError extends ApplicationError {
-  constructor(message: string) {
-    super(message)
-  }
-}
