--- src/app/modules/form/admin-form/fields/fields.types.ts
@@ -1,7 +1,0 @@
-import { Result } from 'neverthrow'
-
-import { IFieldSchema } from '../../../../../types'
-
-import { EditFieldError } from './fields.errors'
-
-export type EditFormFieldResult = Result<IFieldSchema[], EditFieldError>
